backend/requirements.txt
aiosmtplib==5.0.0
alembic==1.16.4
anyio==4.9.0
bcrypt==3.2.0
cffi==1.17.1
click==8.2.2
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
environs==14.3.0
fastapi==0.116.1
fastapi_cors==0.0.6
greenlet==3.2.3
h11==0.16.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==4.0.0
passlib==1.7.4
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==1.10.22
PyJWT==2.6.0
python-dotenv==1.0.0
python-multipart==0.0.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.42
starlette==0.47.3
typing_extensions==4.14.1
uvicorn==0.21.1
